requirements.txt
matplotlib==3.1.0
numpy==1.16.4
scipy==1.2.1
tqdm==4.32.1
tensorflow==1.12.0
pandas==0.24.2
setuptools==41.0.1
Pillow==6.2.1
scikit_learn==0.21.3
